Rewrite rendering to use compiled display lists.
This avoids the need to copy the parameters of each wing across the main bus to the graphics memory on every render. Now when each wing is created, its details are copied to the graphics memory once, and each subsequent render can re-use the data directly from graphics memory.

Or, at least, that is how it is supposed to work. I have no idea if the graphics card drivers actually do such an obvious optimization. But then, the whole point of using a high-level API abstraction is that I do not need to be aware of whether the implementation is currently perfect. All I have to do is make the most efficient use of the abstraction, and future graphics card drivers can provide improvements that my program will transparently use.