Switch buffer data that only lives for one frame to use the STREAM hint.
This is not entirely accurate. The STREAM hint is defined as "defined once and drawn once". Technically it may be possible for the window to be drawn more than once before the animation timer triggers new data generated. But these are only hints anyway, and the standard says that rendering results cannot change based solely on the hint.